java-core-exercises/file-reader/src/main/java/ua/procamp/FileReaders.java
package ua.procamp;

import java.io.IOException;
import java.net.URI;
import java.net.URISyntaxException;
import java.nio.charset.StandardCharsets;
import java.nio.file.Files;
import java.nio.file.Paths;
import java.util.Objects;
import java.util.stream.Stream;

/**
 * {@link FileReaders} privides an API that allow to read whole file into a {@link String} by file name.
 */
public class FileReaders {

    /**
     * Returns a {@link String} that contains whole text from the file specified by name.
     *
     * @param fileName a name of a text file
     * @return string that holds whole file content
     */
    public static String readWholeFile(String fileName) throws URISyntaxException {
        URI path = Objects.requireNonNull(FileReaders.class
                .getClassLoader().getResource(fileName)).toURI();
        StringBuilder contentBuilder = new StringBuilder();
        try (Stream<String> stream = Files.lines(Paths.get(path), StandardCharsets.UTF_8)) {
            stream.forEach(s -> contentBuilder.append(s).append("\n"));
        } catch (IOException e) {
            e.printStackTrace();
        }
        if (contentBuilder.length() > 0) {
            contentBuilder.deleteCharAt(contentBuilder.length() - 1);
        }
        return contentBuilder.toString();
    }
}
